20180126: added css and data to event details page; move elements to its original paths
