+++ src/app/core/core.module.ts
@@ -1,8 +1,9 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { AuthenticationService } from '../services/authentication.service';
-import { MeetupService } from '../services/meetup.service';
 import { HttpClientModule } from '@angular/common/http';
+
+import { AuthenticationService } from './services/authentication.service';
+import { MeetupService } from './services/meetup.service';
 
 @NgModule({
   imports: [
