+++ src/app/core/services/user.service.ts
@@ -1,5 +1,5 @@
 import { Observable } from 'rxjs/Observable';
-import { environment } from './../../environments/environment';
+import { environment } from '../../../environments/environment';
 import { HttpClient, HttpResponse } from '@angular/common/http';
 import { Injectable } from '@angular/core';
